+++ src/components/Edition.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { onMounted, ref, watch } from 'vue'
 
+import SectionBorder from './SectionBorder.vue'
 import SectionTitle from './SectionTitle.vue'
 import Output from './Output.vue'
 import FormInput from './FormInput.vue'
@@ -22,7 +23,7 @@ watch(
 </script>
 
 <template>
-  <section class="mb-5 p-5 border-4 border-gray-500">
+  <SectionBorder>
     <SectionTitle anchor="edition" title="Edition" />
     <div class="flex">
       <form class="w-1/2 pr-2 flex flex-col">
@@ -32,5 +33,5 @@ watch(
         <Output :data="edition"></Output>
       </div>
     </div>
-  </section>
+  </SectionBorder>
 </template>