+++ src/components/GeographicExtent.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { computed, type ComputedRef, ref } from 'vue'
 
+import SectionBorder from './SectionBorder.vue'
 import SectionTitle from './SectionTitle.vue'
 import Output from './Output.vue'
 import FormLabel from './FormLabel.vue'
@@ -37,7 +38,7 @@ let extent: ComputedRef<Extent> = computed(() => {
 </script>
 
 <template>
-  <section class="mb-5 p-5 border-4 border-gray-500">
+  <SectionBorder>
     <SectionTitle anchor="spatial-extent" title="Spatial extent" sub-title="Well-known extents" />
     <div class="flex">
       <form class="w-1/2 pr-2 flex flex-col">
@@ -56,5 +57,5 @@ let extent: ComputedRef<Extent> = computed(() => {
         <Output :data="extent"></Output>
       </div>
     </div>
-  </section>
+  </SectionBorder>
 </template>