+++ src/components/Identifiers.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { ref, watch } from 'vue'
 
+import SectionBorder from './SectionBorder.vue'
 import SectionTitle from './SectionTitle.vue'
 import Output from './Output.vue'
 import IdentifierSelf from './IdentifierSelf.vue'
@@ -37,7 +38,7 @@ watch(
 </script>
 
 <template>
-  <section class="mb-5 p-5 border-4 border-gray-500">
+  <SectionBorder>
     <SectionTitle anchor="identifiers" title="Identifiers" />
     <div class="flex">
       <form class="w-1/2 pr-2 flex flex-col">
@@ -56,5 +57,5 @@ watch(
         <Output :data="identifiers"></Output>
       </div>
     </div>
-  </section>
+  </SectionBorder>
 </template>