+++ src/components/Lineage.vue
@@ -1,11 +1,12 @@
 <script setup lang="ts">
+import SectionBorder from './SectionBorder.vue'
 import SectionTitle from './SectionTitle.vue'
 import Freetext from './Freetext.vue'
 </script>
 
 <template>
-  <section class="mb-5 p-5 border-4 border-gray-500">
+  <SectionBorder>
     <SectionTitle anchor="lineage" title="Lineage" />
     <Freetext input-class="min-h-60" />
-  </section>
+  </SectionBorder>
 </template>