+++ src/components/ResourceType.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { onMounted, ref, watch } from 'vue'
 
+import SectionBorder from './SectionBorder.vue'
 import SectionTitle from './SectionTitle.vue'
 import Output from './Output.vue'
 import FormLabel from './FormLabel.vue'
@@ -24,7 +25,7 @@ watch(
 </script>
 
 <template>
-  <section class="mb-5 p-5 border-4 border-gray-500">
+  <SectionBorder>
     <SectionTitle anchor="resource-type" title="Resource Type" />
     <div class="flex">
       <form class="w-1/2 pr-2 flex flex-col">
@@ -43,5 +44,5 @@ watch(
         <Output :data="resourceType"></Output>
       </div>
     </div>
-  </section>
+  </SectionBorder>
 </template>