+++ src/components/Services.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import servicesData from '../data/services.json'
 
+import SectionBorder from './SectionBorder.vue'
 import SectionTitle from './SectionTitle.vue'
 import Service from './Service.vue'
 
@@ -9,10 +10,10 @@ const serviceSlugs: string[] = Object.keys(servicesData.services).sort((a: strin
 )
 </script>
 <template>
-  <section class="mb-5 p-5 border-4 border-gray-500">
+  <SectionBorder>
     <SectionTitle anchor="services" title="Services" />
     <div class="space-y-4">
       <Service v-for="slug in serviceSlugs" v-bind:key="slug" :slug="slug" />
     </div>
-  </section>
+  </SectionBorder>
 </template>