+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Refactoring prose HTML output sections into a `Prose` component
 - Refactoring text areas into a `FormTextarea` component
 - Refactoring non-radio/checkbox form inputs into a `FormInput` component
+- Refactor section wrappers into a `SectionBorder` component
 
 ## [0.4.0] - 2023-12-27